Updated 04 demo to use code from latest ai demo
