+++ 04-oauth-pivot/src/utils.ts
@@ -78,3 +78,12 @@ export async function fetchUpstreamAuthToken({
   }
   return [accessToken, null]
 }
+
+// Context from the auth process, encrypted & stored in the auth token
+// and provided to the DurableMCP as this.props
+export type Props = {
+  login: string
+  name: string
+  email: string
+  accessToken: string
+}
