+++ 04-oauth-pivot/tsconfig.json
@@ -12,7 +12,7 @@
     /* Specify what module code is generated. */
     "module": "es2022",
     /* Specify how TypeScript looks up a file from a given module specifier. */
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     /* Specify type package names to be included without being referenced in a source file. */
     "types": ["@cloudflare/workers-types/2023-07-01"],
     /* Enable importing .json files */
